Conflicts:
	nsd/Makefile.bsd-wrapper
	nsd/Makefile.in
	nsd/acx_nlnetlabs.m4
	nsd/axfr.c
	nsd/buffer.c
	nsd/compat/b64_pton.c
	nsd/compat/snprintf.c
	nsd/config.h.in
	nsd/configlexer.lex
	nsd/configparser.y
	nsd/configure
	nsd/configure.ac
	nsd/dbaccess.c
	nsd/dbcreate.c
	nsd/difffile.c
	nsd/dname.c
	nsd/dns.c
	nsd/dns.h
	nsd/mkinstalldirs
	nsd/namedb.h
	nsd/nsd-checkconf.8
	nsd/nsd-checkconf.c
	nsd/nsd-notify.8
	nsd/nsd-notify.c
	nsd/nsd-patch.8
	nsd/nsd-patch.c
	nsd/nsd-xfer.8
	nsd/nsd-xfer.c
	nsd/nsd.8
	nsd/nsd.c
	nsd/nsd.conf.5
	nsd/nsd.conf.sample.in
	nsd/nsd.h
	nsd/nsdc.8
	nsd/nsdc.sh.in
	nsd/options.c
	nsd/options.h
	nsd/query.c
	nsd/query.h
	nsd/rdata.c
	nsd/region-allocator.c
	nsd/region-allocator.h
	nsd/server.c
	nsd/tsig-openssl.c
	nsd/tsig-openssl.h
	nsd/tsig.c
	nsd/util.c
	nsd/util.h
	nsd/xfrd-notify.c
	nsd/xfrd-notify.h
	nsd/xfrd-tcp.c
	nsd/xfrd.c
	nsd/xfrd.h
	nsd/zlexer.lex
	nsd/zonec.c

+++ pkg_add/OpenBSD/AddDelete.pm
@@ -1,7 +1,19 @@
 # ex:ts=8 sw=4:
-# $OpenBSD: AddDelete.pm,v 1.15 2010/01/03 09:34:22 espie Exp $
+# $OpenBSD: AddDelete.pm,v 1.38 2010/08/13 11:12:04 espie Exp $
 #
-# Copyright (c) 2007-2009 Marc Espie <espie@openbsd.org>
+# Copyright (c) 2007-2010 Marc Espie <espie@openbsd.org>
+#
+# Permission to use, copy, modify, and distribute this software for any
+# purpose with or without fee is hereby granted, provided that the above
+# copyright notice and this permission notice appear in all copies.
+#
+# THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+# WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+# MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+# ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+# WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+# ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+# OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #
 
 use strict;
@@ -12,72 +24,30 @@ package main;
 our $not;
 
 package OpenBSD::AddDelete;
-use OpenBSD::Getopt;
 use OpenBSD::Error;
 use OpenBSD::Paths;
-
-our $bad = 0;
-our %defines = ();
-our $state;
-
-our ($opt_n, $opt_x, $opt_v, $opt_B, $opt_L, $opt_i, $opt_q, $opt_c, $opt_I, $opt_s);
-$opt_v = 0;
-
-sub handle_options
-{
-	my ($opt_string, $hash, @usage) = @_;
-
-	set_usage(@usage);
-	$state = OpenBSD::State->new;
-	$hash->{h} = sub { Usage(); };
-	$hash->{f} = $hash->{F} = sub { 
-		for my $o (split /\,/o, shift) { 
-			$defines{$o} = 1;
-		}
-	};
-	try {
-		getopts('hciInqvsxB:f:F:L:'.$opt_string, $hash);
-	} catchall {
-		Usage($_);
-	};
-
-	$opt_L = OpenBSD::Paths->localbase unless defined $opt_L;
-
-	if ($opt_s) {
-		$opt_n = 1;
-	}
-	$state->{not} = $opt_n;
-	# XXX RequiredBy
-	$main::not = $opt_n;
-	$state->{defines} = \%defines;
-	$state->{interactive} = $opt_i;
-	$state->{v} = $opt_v;
-	$state->{localbase} = $opt_L;
-	$state->{size_only} = $opt_s;
-	$state->{quick} = $opt_q;
-	$state->{extra} = $opt_c;
-	$state->{dont_run_scripts} = $opt_I;
-}
+use OpenBSD::PackageInfo;
+use OpenBSD::AddCreateDelete;
 
 sub do_the_main_work
 {
-	my $code = shift;
+	my ($self, $state) = @_;
 
-	if ($bad) {
+	if ($state->{bad}) {
 		exit(1);
 	}
 
-	my $handler = sub { my $sig = shift; die "Caught SIG$sig"; };
+	my $handler = sub { $state->fatal("Caught SIG#1", shift); };
 	local $SIG{'INT'} = $handler;
 	local $SIG{'QUIT'} = $handler;
 	local $SIG{'HUP'} = $handler;
 	local $SIG{'KILL'} = $handler;
 	local $SIG{'TERM'} = $handler;
 
-	if ($state->{defines}->{debug}) {
-		&$code;
-	} else { 
-		eval { &$code; };
+	if ($state->defines('debug')) {
+		$self->main($state);
+	} else {
+		eval { $self->main($state); };
 	}
 	my $dielater = $@;
 	return $dielater;
@@ -85,39 +55,57 @@ sub do_the_main_work
 
 sub framework
 {
-	my $code = shift;
-	try {
-		lock_db($opt_n) unless $state->{defines}->{nolock};
-		$state->setup_progressmeter($opt_x);
+	my ($self, $state) = @_;
+
+	my $do = sub {
+		lock_db($state->{not}) unless $state->defines('nolock');
 		$state->check_root;
-		process_parameters();
-		my $dielater = do_the_main_work($code);
+		$self->process_parameters($state);
+		my $dielater = $self->do_the_main_work($state);
 		# cleanup various things
 		$state->{recorder}->cleanup($state);
 		OpenBSD::PackingElement::Lib::ensure_ldconfig($state);
 		OpenBSD::PackingElement::Fontdir::finish_fontdirs($state);
 		$state->progress->clear;
 		$state->log->dump;
-		finish_display();
-		if ($state->verbose >= 2 || $opt_s || 
-		    $state->{defines}->{tally}) {
+		$self->finish_display($state);
+		if ($state->verbose >= 2 || $state->{size_only} ||
+		    $state->defines('tally')) {
 			$state->vstat->tally;
 		}
 		# show any error, and show why we died...
 		rethrow $dielater;
-	} catch {
-		print STDERR "$0: $_\n";
-		if ($_ =~ m/^Caught SIG(\w+)/o) {
-			kill $1, $$;
-		}
-		exit(1);
 	};
+	if ($state->defines('debug')) {
+		&$do;
+	} else {
+		try {
+			&$do;
+		} catch {
+			print STDERR "$0: $_\n";
+			OpenBSD::Handler->reset;
+			if ($_ =~ m/^Caught SIG(\w+)/o) {
+				kill $1, $$;
+			}
+			exit(1);
+		};
+	}
 
-	if ($bad) {
+	if ($state->{bad}) {
 		exit(1);
 	}
 }
 
+sub parse_and_run
+{
+	my ($self, $cmd) = @_;
+
+	my $state = $self->new_state($cmd);
+	$state->handle_options;
+	local $SIG{'INFO'} = sub { $state->status->print($state); };
+
+	$self->framework($state);
+}
 
 package OpenBSD::SharedItemsRecorder;
 sub new
@@ -142,74 +130,71 @@ sub cleanup
 	OpenBSD::SharedItems::cleanup($self, $state);
 }
 
-package OpenBSD::Log;
-use OpenBSD::Error;
-our @ISA = qw(OpenBSD::Error);
-
-sub set_context
-{
-	&OpenBSD::Error::set_pkgname;
-}
-
-sub dump
-{
-	&OpenBSD::Error::delayed_output;
-}
-
-
-package OpenBSD::UI;
-use OpenBSD::Error;
+package OpenBSD::AddDelete::State;
 use OpenBSD::Vstat;
+use OpenBSD::Log;
+our @ISA = qw(OpenBSD::AddCreateDelete::State);
 
-sub new
+sub handle_options
 {
-	my $class = shift;
-	my $o = bless {}, $class;
-	$o->init(@_);
-	return $o;
+	my ($state, $opt_string, @usage) = @_;
+
+	# backward compatibility
+	$state->{opt}{F} = sub {
+		for my $o (split /\,/o, shift) {
+			$state->{subst}->add($o, 1);
+		}
+	};
+	$state->{no_exports} = 1;
+	$state->SUPER::handle_options($opt_string.'ciInqsB:F:', @usage);
+
+	if ($state->opt('s')) {
+		$state->{not} = 1;
+	}
+	# XXX RequiredBy
+	$main::not = $state->{not};
+	$state->{interactive} = $state->opt('i');
+	$state->{localbase} = $state->opt('L') // OpenBSD::Paths->localbase;
+	$state->{size_only} = $state->opt('s');
+	$state->{quick} = $state->opt('q') || $state->config->istrue("nochecksum");
+	$state->{extra} = $state->opt('c');
+	$state->{dont_run_scripts} = $state->opt('I');
+	$ENV{'PKG_DELETE_EXTRA'} = $state->{extra} ? "Yes" : "No";
 }
 
 sub init
 {
 	my $self = shift;
-	$self->{l} = OpenBSD::Log->new;
+	$self->{l} = OpenBSD::Log->new($self);
 	$self->{vstat} = OpenBSD::Vstat->new($self);
-	$self->{progressmeter} = bless {}, "OpenBSD::StubProgress";
 	$self->{status} = OpenBSD::Status->new;
 	$self->{recorder} = OpenBSD::SharedItemsRecorder->new;
 	$self->{v} = 0;
+	$self->{wantntogo} = $self->config->istrue("ntogo");
+	$self->SUPER::init(@_);
+	$self->{export_level}++;
 }
 
 sub ntogo
 {
 	my ($self, $offset) = @_;
 
-	return $self->progress->ntogo($self->todo, $offset);
+	return $self->{wantntogo} ?
+	    $self->progress->ntogo($self, $offset) :
+	    $self->f("ok");
 }
 
 sub ntogo_string
 {
-	my ($self, $todo, $offset) = @_;
+	my ($self, $offset) = @_;
 
-	$todo //= 0;
-	$offset //= 0;
-	$todo += $offset;
-
-	if ($todo > 0) {
-		return " ($todo to go)";
-	} else {
-		return "";
-	}
-}
-
-sub verbose
-{
-	return shift->{v};
+	return $self->todo($offset // 0);
 }
 
 sub vstat
 {
-	return shift->{vstat};
+	my $self = shift;
+	return $self->{vstat};
 }
 
 sub log
@@ -218,79 +203,37 @@ sub log
 	if (@_ == 0) {
 		return $self->{l};
 	} else {
-		$self->{l}->print(@_);
+		$self->{l}->say(@_);
 	}
-}
-
-sub print
-{
-	my $self = shift;
-	$self->progress->print(@_);
-}
-
-sub say
-{
-	my $self = shift;
-	$self->progress->print(@_, "\n");
-}
-
-sub errprint
-{
-	my $self = shift;
-	$self->progress->errprint(@_);
-}
-
-sub errsay
-{
-	my $self = shift;
-	$self->progress->errprint(@_, "\n");
-}
-
-sub progress
-{
-	my $self = shift;
-	return $self->{progressmeter};
 }
 
 sub vsystem
 {
 	my $self = shift;
-	$self->progress->clear;
-	OpenBSD::Error::VSystem($self->verbose >= 2, @_);
+	my $verbose = $self;
+	if ($self->verbose < 2) {
+		$self->system(@_);
+	} else {
+		$self->verbose_system(@_);
+	}
 }
 
 sub system
 {
 	my $self = shift;
 	$self->progress->clear;
-	OpenBSD::Error::System(@_);
-}
-
-sub unlink
-{
-	my $self = shift;
-	$self->progress->clear;
-	OpenBSD::Error::Unlink(@_);
-}
-
-# we always have a progressmeter we can print to...
-sub setup_progressmeter
-{
-	my ($self, $opt_x) = @_;
-	if (!$opt_x) {
-		require OpenBSD::ProgressMeter;
-		$self->{progressmeter} = OpenBSD::ProgressMeter->new;
-	}
+	$self->SUPER::system(@_);
 }
 
 sub check_root
 {
 	my $state = shift;
-	if ($< && !$state->{defines}->{nonroot}) {
+	if ($< && !$state->defines('nonroot')) {
 		if ($state->{not}) {
-			$state->errsay("$0 should be run as root") if $state->verbose;
+			$state->errsay("#1 should be run as root",
+			    $state->{cmd}) if $state->verbose;
 		} else {
-			Fatal "$0 must be run as root";
+			$state->fatal("#1 must be run as root", $state->{cmd});
 		}
 	}
 }
@@ -299,7 +242,7 @@ sub choose_location
 {
 	my ($state, $name, $list, $is_quirks) = @_;
 	if (@$list == 0) {
-		$state->errsay("Can't find $name") unless $is_quirks;
+		$state->errsay("Can't find #1", $name) unless $is_quirks;
 		return undef;
 	} elsif (@$list == 1) {
 		return $list->[0];
@@ -311,10 +254,11 @@ sub choose_location
 
 		$h{'<None>'} = undef;
 		$state->progress->clear;
-		my $result = OpenBSD::Interactive::ask_list("Ambiguous: choose package for $name", 1, sort keys %h);
+		my $result = $state->ask_list("Ambiguous: choose package for $name", 1, sort keys %h);
 		return $h{$result};
 	} else {
-		$state->errsay("Ambiguous: $name could be ", join(' ', keys %h));
+		$state->errsay("Ambiguous: #1 could be #2",
+		    $name, join(' ', keys %h));
 		return undef;
 	}
 }
@@ -328,11 +272,68 @@ sub confirm
 	return OpenBSD::Interactive::confirm($prompt, $default);
 }
 
+sub ask_list
+{
+	my ($state, $prompt, $interactive, @values) = @_;
+
+	require OpenBSD::Interactive;
+	return OpenBSD::Interactive::ask_list($prompt, $interactive, @values);
+}
+
 sub status
 {
 	my $self = shift;
 
 	return $self->{status};
+}
+
+sub defines
+{
+	my ($self, $k) = @_;
+	return $self->{subst}->value($k);
+}
+
+sub updateset
+{
+	my $self = shift;
+	require OpenBSD::UpdateSet;
+
+	return OpenBSD::UpdateSet->new($self);
+}
+
+sub updateset_with_new
+{
+	my ($self, $pkgname) = @_;
+
+	return $self->updateset->add_newer(
+	    OpenBSD::Handle->create_new($pkgname));
+}
+
+sub updateset_from_location
+{
+	my ($self, $location) = @_;
+
+	return $self->updateset->add_newer(
+	    OpenBSD::Handle->from_location($location));
+}
+
+# those are required for makewhatis integration
+sub picky
+{
+	return shift->{picky};
+}
+
+sub testmode
+{
+	return shift->{testmode};
+}
+
+sub check_dir
+{
+	my ($self, $dir) = @_;
+	unless (-d $dir) {
+	    $self->fatal("#1: #2 is not a directory", $0, $dir);
+	}
 }
 
 # the object that gets displayed during status updates
@@ -353,8 +354,8 @@ sub print
 		$object = "Parameters";
 	}
 
-	$state->say($what." ".$object.$state->ntogo_string($state->todo));
-	if ($state->{defines}->{carp}) {
+	$state->say($what." #1 (#2)", $object, $state->ntogo_string);
+	if ($state->defines('carp')) {
 		require Carp;
 		Carp::cluck("currently here");
 	}
@@ -388,76 +389,6 @@ sub new
 	my $class = shift;
 
 	bless {}, $class;
-}
-
-# stub class when no actual progressmeter that still prints out.
-package OpenBSD::StubProgress;
-sub clear {}
-
-sub show {}
-
-sub message {}
-
-sub next {}
-
-sub set_header {}
-
-sub print
-{
-	shift;
-	print @_;
-}
-
-sub ntogo
-{
-	return "";
-}
-
-sub errprint
-{
-	shift;
-	print STDERR @_;
-}
-
-package OpenBSD::PackingList;
-sub compute_size
-{
-	my $plist = shift;
-	my $totsize = 0;
-	$plist->visit('compute_size', \$totsize);
-	$totsize = 1 if $totsize == 0;
-	$plist->{totsize} = $totsize;
-}
-
-package OpenBSD::PackingElement;
-sub mark_progress
-{
-}
-
-sub compute_size
-{
-}
-
-package OpenBSD::PackingElement::FileBase;
-sub mark_progress
-{
-	my ($self, $progress, $donesize, $totsize) = @_;
-	return unless defined $self->{size};
-	$$donesize += $self->{size};
-	$progress->show($$donesize, $totsize);
-}
-
-sub compute_size
-{
-	my ($self, $totsize) = @_;
-
-	$$totsize += $self->{size} if defined $self->{size};
-}
-
-package OpenBSD::PackingElement::Sample;
-sub compute_size
-{
-	&OpenBSD::PackingElement::FileBase::compute_size;
 }
 
 1;
